FiX Counsel fuctions

+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/counsel/service/CounselService.java
@@ -1,5 +1,6 @@
 package com.EmperorPenguin.SangmyungBank.counsel.service;
 
+import com.EmperorPenguin.SangmyungBank.baseUtil.config.DateConfig;
 import com.EmperorPenguin.SangmyungBank.baseUtil.exception.CounselException;
 import com.EmperorPenguin.SangmyungBank.baseUtil.exception.ExceptionMessages;
 import com.EmperorPenguin.SangmyungBank.counsel.dto.CounselCreateReq;
@@ -7,6 +8,7 @@ import com.EmperorPenguin.SangmyungBank.counsel.dto.CounselRequestRes;
 import com.EmperorPenguin.SangmyungBank.counsel.dto.CounselUpdateReq;
 import com.EmperorPenguin.SangmyungBank.counsel.entity.Counsel;
 import com.EmperorPenguin.SangmyungBank.counsel.repository.CounselRepository;
+import com.EmperorPenguin.SangmyungBank.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,18 +19,24 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class CounselService {
+
     private final CounselRepository counselRepository;
+    private final UserRepository userRepository;
 
     @Transactional
     public void createCounsel(CounselCreateReq counselCreateReq) {
-        if (counselRepository.findByTitle(counselCreateReq.getTitle()).isPresent()) {
-            throw new CounselException(ExceptionMessages.ERROR_EVENT_EXIST);
+        String loginId = counselCreateReq.getLoginId();
+        // 먼저 해당 사용자가 있는 지 검증
+        if(!userRepository.existsByLoginId(loginId)){
+            throw new CounselException(ExceptionMessages.ERROR_USER_NOT_FOUND);
         }
         try {
-            counselRepository.save(counselCreateReq.toEntity());
+            counselRepository.save(counselCreateReq.toEntity(
+                    userRepository.findByLoginId(loginId).get())
+            );
         } catch (Exception e) {
             e.printStackTrace();
-            throw new CounselException("이벤트 생성에 실패했습니다.");
+            throw new CounselException("상담글 생성에 실패했습니다.");
         }
     }
 
@@ -55,8 +63,15 @@ public class CounselService {
         if(!counselRepository.existsById(counselUpdateReq.getId())){
             throw new CounselException(ExceptionMessages.ERROR_EVENT_NOT_EXIST);
         }
+        Counsel counsel = counselRepository.getById(counselUpdateReq.getId());
+        if(!counsel.getUserId().getLoginId().equals(counselUpdateReq.getLoginId())){
+            throw new CounselException(ExceptionMessages.ERROR_COUNSEL_UNAUTHORIZED_ACCESS);
+        }
         try {
-            counselRepository.updateCounsel(counselUpdateReq.getId(),counselUpdateReq.getTitle(),counselUpdateReq.getContent());
+            counselRepository.updateCounsel(counselUpdateReq.getId(),
+                    counselUpdateReq.getTitle(),
+                    counselUpdateReq.getContent(),
+                    new DateConfig().getDateTime());
         }catch (Exception e){
             e.printStackTrace();
             throw new CounselException("상담글 업데이트에 실패했습니다.");
@@ -64,11 +79,15 @@ public class CounselService {
     }
 
     @Transactional
-    public void deleteCounsel(Long id) {
+    public void deleteCounsel(Long id, String loginId) {
         if(!counselRepository.existsById(id)){
             throw new CounselException(ExceptionMessages.ERROR_COUNSEL_NOT_EXIST);
         }
-        try {
+        Counsel counsel = counselRepository.getById(id);
+        if(!counsel.getUserId().getLoginId().equals(loginId)){
+            throw new CounselException(ExceptionMessages.ERROR_COUNSEL_UNAUTHORIZED_ACCESS);
+        }
+        try{
             counselRepository.deleteById(id);
         }catch (Exception e){
             e.printStackTrace();
